experimental implementation of Go codegen for JSON Schema
It results in the same Go codegen output as the previous library we used for the same task (github.com/a-h/generate), except for the top doc comment and one difference in field sorting (sorting by the JSON Schema property name vs. the Go name), both of which are insignificant.

It will let us improve our JSON config with defaults and union types (and other things).