Fix output for nullable types
When a type is declared nullable, we wouldn't need to care about it in Go for arrays and objects, but currently we fall back to just `interface{}`. This fixes it, by accounting for nullability in the type checks in the generator.

Before, `type: [array, null]` would become `interface{}`, now it becomes the same generated type it would when it was non-nullable.