compiler: handle references to files in the same directory
I was attempting to factor out some common schema properties into a
shared file. However, I was getting an error like the below

  compilation error: failed to resolve $ref:
  "ratelimit.schema.json#" (dereferenced to "/ratelimit.schema.json").

This is due to our use of ResolveReference which would turnñ a
relative reference into an absolute, and then fail to find it. This
fixes that.

Test Plan: used this version in my in progress PR in Sourcegraph and the
compiler worked.