fix handling of additionalProperties when there are >1 explicit properties
When an object type had 2 or more explicitly defined properties, the generated code for the Go type's UnmarshalJSON method was syntactically invalid due to a whitespace bug.

Also, the test for the object-with-props test case was not being executed because it was underneath a directory named `testdata`, so it was moved to the `compiler` package.

Finally, there are a few other optimizations to the generated code.